Added a "long_id" field to the SDF block type.
If a block ID is longer than "c_id_length" then it will be
truncated. Once this occurs, it is possible that the shortened
ID string does not contain enough information to find it again.
With this change, the original string is stored in the block
structure so that up until the SDF file is closed the block can
still be retrieved.