Added safe_copy_unique_id to the SDF routines.
This routine does the job of testing that the block ID string
is unique within the file and auto-generating a new unique ID
if necessary. It is similar to the get_unique_id() routine it
has replaced, except that it is now passed the block so that it
can change other attributes if necessary.