Added routines for reading global station information.
This adds a couple of helper routines which returns the global
station headers.
When writing output to the station files, each time the list of
stations tracked changes, a new block is written whose header
contains the ids and other information about the new list of
stations. When reading the station file, it is then necessary
to construct a list of all station ids used in the file and for
each block, determine where the station ids fit into this global
list.
The routine sdf_get_all_stations() returns the global list of
station ids. The routine sdf_read_station_info_arrays_all()
returns a global version of the station_move and station_nvars
arrays.