Reduce module USE's to make ifort-13.x happy.
Recent versions of the Intel Fortran compiler take hours to
compile the code if there are recursively included modules.
Although this is quite clearly a compiler bug, it is not too
painful to work around.
It may turn out that this causes issues with other compilers and
may need to be reverted.