Hack to work around issues on filesystems without flock().
If a filesystem is mounted without support for the flock()
system call then non-interleaved parallel writes will fail.
This patch enables a ROMIO hint which is known to work around
the issue on some systems.
It is not a true fix because the underlying problem is with the
filesystem and not EPOCH. It is a bit of a cludge and is not
guaranteed to work on all MPI implementations.