Fixed some CMake string variables.
In CMake the syntax:
   "string 1" "string 2"
does not concatenate strings, like it does in C or Python. Instead it is
a list of strings, and appears as
   "string 1 ; string 2"
when the variable is actually used.

This change effectively reverts a previous commit that tidied up lines
longer than 80 columns. I think the choice is either to have long lines
or multiple commands like:
    set(VAR "${VAR} extra_chars")

CMake v3.0 and later allows '\' to split a string across several lines,
but this is not compatible with earlier versions.