Add include directory to simplify external Makefiles
The typical method of compiling against SDF is to add the include
directory from this repository to the list of include paths. This can
cause problems when the directory does not exist. There are other
possible approaches to fixing the problem but just ensuring that the
directory always exists seems to be the easiest.